Clean-up azure storage client and controller
